refactor(tests): fix additional clippy warnings
This fixes warnings that were reported using
[bacon](https://github.com/Canop/bacon/tree/main)'s `clippy-all` job.

All of the warnings were for
<https://rust-lang.github.io/rust-clippy/master/index.html#needless_borrow>

I ran into this by accident and figured it might be a good idea to fix
them. I researched this a bit and think in practice they will have
little to no performance impact - this seems to be motivated by
simplicity and readability.